pyproject.toml
[tool.poetry]
name = "dms-python-cookbook"
version = "0.1.0"
description = ""
authors = ["ER-WagnerMatth <matthias.wagner@stadt.erlangen.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
dotenv = "^0.9.9"
requests = "^2.32.3"
pytest = "^8.3.5"
fastapi = "^0.115.12"
importlib = "^1.0.4"
uvicorn = "^0.34.3"
sqlalchemy = "^2.0.41"
psycopg = {extras = ["binary"], version = "^3.2.9"}
argparse = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
